traderbot/requirements.txt
--extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments
pandas
python-dotenv
pyyaml
requests
pyTelegramBotAPI
